ch06/requirements.txt
Flask
python-dotenv
selenium
line-bot-sdk
matplotlib
pandas
mplfinance
TA-Lib
